[JDR-Main-Calculator-Hook] Add calculator hook and rebuild main-button component, dispatch calculate when import pz
